Catch Throwable in ThrowableExtendedStackTraceRenderer#loadClass (#4028)
- Changed catch (Exception) to catch (Throwable) in loadClass method
- Moved tests into ThrowablePatternConverterTest.StackTraceTest
- Removed standalone ThrowableExtendedStackTraceRendererTest
- Updated changelog description and issue link per review feedback
