+++ log4j-core/src/main/java/org/apache/logging/log4j/core/pattern/ThrowableExtendedStackTraceRenderer.java
@@ -181,7 +181,7 @@ final class ThrowableExtendedStackTraceRenderer
                     if (clazz != null) {
                         return clazz;
                     }
-                } catch (final Exception | LinkageError ignored) {
+                } catch (final Throwable ignored) {
                     // Do nothing
                 }
             }
