+++ src/changelog/.2.x.x/4049_fix_catch_LinkageError_in_ThrowableExtendedStackTraceRenderer.xml
@@ -4,6 +4,6 @@
        xsi:schemaLocation="https://logging.apache.org/xml/ns
        https://logging.apache.org/xml/ns/log4j-changelog-0.xsd"
        type="fixed">
-    <issue id="4028" link="https://github.com/apache/logging-log4j2/issues/4028"/>
-    <description format="asciidoc">Catch `LinkageError` in `ThrowableExtendedStackTraceRenderer#loadClass` to prevent `NoClassDefFoundError` from breaking logging</description>
+    <issue id="4049" link="https://github.com/apache/logging-log4j2/pull/4049"/>
+    <description format="asciidoc">Ensure all `Throwable`s are handled while rendering stack traces</description>
 </entry>
